requirements.txt
pytest
requests
playwright
allure-pytest
pandas
google-generativeai
pytest-xdist
pytest-rerunfailures
